Fix dependencies, more comments
